refactor: button 관련 상수 분리
