+++ frontend/src/components/Buttons/IconToggleButton.tsx
@@ -1,10 +1,15 @@
-import { ButtonSize } from './Button';
+import type { ButtonColor, IconColor } from '@constants/colors';
+import type { ButtonSize } from '@constants/sizes';
+
 import IconButton from './IconButton';
 
 interface IconToggleButtonProps {
   activeIcon: string;
-  disabledIcon: string;
+  inactiveIcon: string;
+  iconColor?: IconColor;
   iconSize?: number;
+  buttonActiveColor?: ButtonColor;
+  buttonInactiveColor?: ButtonColor;
   buttonSize?: ButtonSize;
   onClick?: () => void;
   active: boolean;
@@ -12,21 +17,23 @@ interface IconToggleButtonProps {
 
 export default function IconToggleButton({
   activeIcon,
-  disabledIcon,
+  inactiveIcon,
+  iconColor = 'textWhite',
   iconSize,
+  buttonActiveColor = 'active',
+  buttonInactiveColor = 'cancel',
   buttonSize,
   onClick,
   active,
 }: IconToggleButtonProps) {
   return (
     <IconButton
-      icon={active ? activeIcon : disabledIcon}
+      icon={active ? activeIcon : inactiveIcon}
       iconSize={iconSize}
-      iconColor={'textWhite'}
+      iconColor={iconColor}
       buttonSize={buttonSize}
-      buttonColor={active ? 'active' : 'cancel'}
+      buttonColor={active ? buttonActiveColor : buttonInactiveColor}
       onClick={onClick}
-      circle
     />
   );
 }
